实现增加账户接口

+++ src/main/java/com/augustxun/safe/model/vo/CheckingAccountVO.java
@@ -12,7 +12,7 @@ import java.util.Date;
 public class CheckingAccountVO {
 
     private static final long serialVersionUID = 1L;
-    private Integer acctNo;
+    private Long acctNo;
     private String acctName;
     private String zipcode;
     private String unit;
