+++ src/main/java/com/augustxun/safe/model/vo/PersonalLoanVO.java
@@ -10,7 +10,7 @@ import java.util.Date;
 @Data
 public class PersonalLoanVO implements Serializable {
     private static final long serialVersionUID = 1L;
-    private Integer acctNo;
+    private Long acctNo;
     private String acctName;
     private String zipcode;
     private String unit;
