+++ src/main/java/com/augustxun/safe/model/vo/SavingsAccountVO.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @Data
 public class SavingsAccountVO {
     private static final long serialVersionUID = 1L;
-    private Integer acctNo;
+    private Long acctNo;
     private String acctName;
     private String zipcode;
     private String unit;
